Add extension comments and regroup them
Add doc comments on all extensions.

Merge some related extensions into larger groups and add comments that
specifically describe differences between them.
