+++ lib/src/scan.dart
@@ -4,14 +4,15 @@
 
 import 'dart:async';
 
+/// A utility similar to [fold] which emits intermediate accumulations.
 extension Scan<T> on Stream<T> {
   /// Like [fold], but instead of producing a single value it yields each
   /// intermediate accumulation.
   ///
   /// If [combine] returns a Future it will not be called again for subsequent
-  /// events from the source until it completes, therefor the combine callback
-  /// is always called for elements in order, and the result stream always
-  /// maintains the same order as the original.
+  /// events from the source until it completes, therefore [combine] is always
+  /// called for elements in order, and the result stream always maintains the
+  /// same order as the original.
   Stream<S> scan<S>(
       S initialValue, FutureOr<S> combine(S previousValue, T element)) {
     var accumulated = initialValue;
