+++ lib/src/switch.dart
@@ -4,6 +4,8 @@
 
 import 'dart:async';
 
+/// A utility to take events from the most recent sub stream returned by a
+/// callback.
 extension Switch<T> on Stream<T> {
   /// Maps events to a Stream and emits values from the most recently created
   /// Stream.
@@ -18,6 +20,7 @@ extension Switch<T> on Stream<T> {
   }
 }
 
+/// A utility to take events from the most recent sub stream.
 extension SwitchLatest<T> on Stream<Stream<T>> {
   /// Emits values from the most recently emitted Stream.
   ///
@@ -32,11 +35,11 @@ extension SwitchLatest<T> on Stream<Stream<T>> {
 /// Maps events to a Stream and emits values from the most recently created
 /// Stream.
 ///
-/// When the source emits a value it will be converted to a [Stream] using [map]
-/// and the output will switch to emitting events from that result.
+/// When the source emits a value it will be converted to a [Stream] using
+/// [convert] and the output will switch to emitting events from that result.
 ///
 /// If the source stream is a broadcast stream, the result stream will be as
-/// well, regardless of the types of the streams produced by [map].
+/// well, regardless of the types of the streams produced by [convert].
 @Deprecated('Use the extension instead')
 StreamTransformer<S, T> switchMap<S, T>(Stream<T> convert(S event)) =>
     StreamTransformer.fromBind(
