+++ lib/src/take_until.dart
@@ -4,8 +4,9 @@
 
 import 'dart:async';
 
+/// A utility to end a stream based on an external trigger.
 extension TakeUntil<T> on Stream<T> {
-  /// Returns a stram which emits values from the source stream until [trigger]
+  /// Returns a stream which emits values from the source stream until [trigger]
   /// fires.
   ///
   /// Completing [trigger] differs from canceling a subscription in that values
